cv_to_txt/requirements.txt
pdfplumber
PyMuPDF
python-docx
docx2txt
beautifulsoup4
requests
python-pptx
striprtf